backend_django/requirements.txt
# Django & DRF
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
django-environ==0.11.2

# CORS
django-cors-headers==4.3.1

# Image Handling
Pillow==10.1.0

# Security & Utils
django-filter==23.5
python-decouple==3.8
gunicorn==21.2.0

# Excel Processing
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.2

# Development
django-debug-toolbar==4.2.0

# Deployment (optional)
whitenoise==6.6.0
